add flexible unit conversion system

+++ app/models/inventory.py
@@ -1,6 +1,6 @@
 # app/models/inventory.py
 
-from sqlalchemy import CheckConstraint, Column, Index, Integer, Date, ForeignKey
+from sqlalchemy import CheckConstraint, Column, Index, Integer, Date, ForeignKey, Numeric
 from sqlalchemy.orm import relationship
 
 from app.database import Base
@@ -10,9 +10,29 @@ class Inventory(Base):
     __tablename__ = "inventory"
 
     id = Column(Integer, primary_key=True, index=True)
-    product_id = Column(Integer, ForeignKey("products.id", ondelete="CASCADE"), nullable=False, unique=True, index=True)
-    quantity_available = Column(Integer, nullable=False)
+
+    product_id = Column(
+        Integer,
+        ForeignKey("products.id", ondelete="CASCADE"),
+        nullable=False,
+        unique=True,
+        index=True,
+    )
+
+    # CHANGED: Integer → Numeric
+    # Inventory must support decimal quantities because
+    # unit conversion may result in fractional base units.
+    #
+    # Example:
+    # Base unit = Kongo
+    # 3 Cups sold
+    # 3 Cups = 0.6 Kongo
+    #
+    # Inventory becomes 49.4 Kongos
+    quantity_available = Column(Numeric(14, 4), nullable=False)
+
     low_stock_threshold = Column(Integer, nullable=False, default=5)
+
     expiry_date = Column(Date, nullable=True)
 
     product = relationship("Product", back_populates="inventory")
